+++ packages/contracts/test/voting/allowlist-voting-setup.ts
@@ -2,7 +2,7 @@ import {expect} from 'chai';
 import {BigNumberish} from 'ethers';
 import {ethers} from 'hardhat';
 
-import {AllowlistVotingSetupV1} from '../../typechain';
+import {AllowlistVotingSetup} from '../../typechain';
 import {deployNewDAO} from '../test-utils/dao';
 
 enum Op {
@@ -17,12 +17,10 @@ const AddressZero = ethers.constants.AddressZero;
 const EMPTY_DATA = '0x';
 const EMPTY_BYTES_32 = `0x${'00'.repeat(32)}`;
 
-// TODO 1. add type GRANT/REVOKE check in permissions
-// TODO 2. in order to detect encode abi for deploy/update, use deployABI/updateABI
 describe('AllowlistVotingSetup', function () {
   let ownerAddress: string;
   let signers: any;
-  let allowlistVotingSetupV1: AllowlistVotingSetupV1;
+  let allowlistVotingSetup: AllowlistVotingSetup;
   let implementationAddress: string;
   let targetDao: any;
 
@@ -32,25 +30,22 @@ describe('AllowlistVotingSetup', function () {
     targetDao = await deployNewDAO(ownerAddress);
 
     const AllowlistVotingSetup = await ethers.getContractFactory(
-      'AllowlistVotingSetupV1'
+      'AllowlistVotingSetup'
     );
-    allowlistVotingSetupV1 = await AllowlistVotingSetup.deploy();
+    allowlistVotingSetup = await AllowlistVotingSetup.deploy();
 
     implementationAddress =
-      await allowlistVotingSetupV1.getImplementationAddress();
+      await allowlistVotingSetup.getImplementationAddress();
   });
 
   describe('prepareInstallation', async () => {
     it('fails if data is empty', async () => {
       await expect(
-        allowlistVotingSetupV1.prepareInstallation(
-          targetDao.address,
-          EMPTY_DATA
-        )
+        allowlistVotingSetup.prepareInstallation(targetDao.address, EMPTY_DATA)
       ).to.be.reverted;
 
       await expect(
-        allowlistVotingSetupV1.prepareInstallation(
+        allowlistVotingSetup.prepareInstallation(
           targetDao.address,
           EMPTY_BYTES_32
         )